test: add deterministic tests/ discovery + --only; extract workable (#1440)
